outsourced Model and Dataset definition and removed typo
