+++ src/audio_to_image_model/audio_to_image_model_def.py
@@ -1,0 +1,57 @@
+import torch.nn as nn
+
+
+# Model Definition
+class CustomCNN(nn.Module):
+    def __init__(self, num_classes=264):
+        super(CustomCNN, self).__init__()
+
+        self.conv1 = nn.Sequential(
+            nn.Conv2d(1, 64, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1)),
+            nn.ReLU(),
+            nn.MaxPool2d(kernel_size=(2, 2), stride=(2, 2))
+        )
+
+        self.conv2 = nn.Sequential(
+            nn.Conv2d(64, 128, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1)),
+            nn.ReLU(),
+            nn.MaxPool2d(kernel_size=(2, 2), stride=(2, 2))
+        )
+
+        self.conv3 = nn.Sequential(
+            nn.Conv2d(128, 256, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1)),
+            nn.ReLU(),
+            nn.MaxPool2d(kernel_size=(2, 2), stride=(2, 2))
+        )
+
+        self.conv4 = nn.Sequential(
+            nn.Conv2d(256, 512, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1)),
+            nn.ReLU(),
+            nn.MaxPool2d(kernel_size=(2, 2), stride=(2, 2))
+        )
+
+        self.conv5 = nn.Sequential(
+            nn.Conv2d(512, 1024, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1)),
+            nn.ReLU(),
+            nn.AdaptiveMaxPool2d((1, 6))
+        )
+
+        self.fc1 = nn.Linear(1024 * 6, 2048)
+        self.relu = nn.ReLU()
+        self.fc2 = nn.Linear(2048, num_classes)
+        self.sigmoid = nn.Sigmoid()
+
+    def forward(self, x):
+        x = self.conv1(x)
+        x = self.conv2(x)
+        x = self.conv3(x)
+        x = self.conv4(x)
+        x = self.conv5(x)
+
+        x = x.view(x.size(0), -1)
+        x = self.fc1(x)
+        x = self.relu(x)
+        x = self.fc2(x)
+        x = self.sigmoid(x)
+
+        return x
